requirements.txt
argparse
pylint
pytest
click
pytest-cov
black
tweepy
pandas
numpy
python-dotenv